refactor[service]: Restructure endpoints and implement DTO pattern
- Update CQRS handlers to map models to DTOs
- Remove unused dependencies and redundant configurations
- Consolidate single-item Get endpoints into list endpoints

+++ service/Models/Category.cs
@@ -4,7 +4,7 @@ public class Category
 {
     public int Id { get; set; }
     public string Name { get; set; } = string.Empty;
-    public decimal Target { get; set; }
+    public double Target { get; set; }
     
     public ICollection<Stock> Stocks { get; set; } = new List<Stock>();
 }
